+++ src/Demo/DemoModule.php
@@ -20,7 +20,7 @@ final class DemoModule implements IModule
     /**
      * @inheritdoc
      */
-    public function build(IApplicationBuilder $appBuilder): void
+    public function configure(IApplicationBuilder $appBuilder): void
     {
         $this->withBinders($appBuilder, new UserServiceBinder())
             ->withProblemDetails(