+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- Updated Composer scripts to not run `php aphiria framework:flushcaches` anymore after `composer install` and `composer update`
+- Updated Composer scripts to not run `php aphiria framework:flushcaches` anymore after `composer install` and `composer update` ([#18](https://github.com/aphiria/app/pull/18))
 
 ## [v1.0.0-alpha2](https://github.com/aphiria/app/compare/v1.0.0-alpha1...v1.0.0-alpha2) (2021-2-15)